requirements.txt
pandas
streamlit
scikit-learn
Pillow
joblib
plotly
pickle